+++ CHANGELOG.md
@@ -6,3 +6,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Issue Fixes
 * [#13](https://github.com/upfrontsystems/izinto-angular/issues/13) add variables to dashboard
+* [#14](https://github.com/upfrontsystems/izinto-angular/issues/14) add collections
+* [#14](https://github.com/upfrontsystems/izinto-angular/issues/14) add dashboards in collections reorder
+* [#14](https://github.com/upfrontsystems/izinto-angular/issues/14) add pasting dashboards and collections
+